abipkgdiff: Avoid wrongly considering binaries being added/removed
When the layout of binaries in a package changes across versions of
the package, abipkgdiff can wrongly consider one particular binary to
be removed from the old package and then added again in the new one,
even though the SONAME (when present) has not changed.

This patch makes it so that only the SONAME (when present) is used as
a key for the binary.  That fixes the issue.

	* tools/abipkgdiff.cc (create_maps_of_package_set_content): Use
	the SONAME when present as the key of the binary.
	* tests/data/test-diff-pkg/dbus-glib-0.104-3.fc23.x86_64--dbus-glib-0.104-3.fc23.armv7hl-report-0.txt: Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>