Bug 31642 - Don't forget filtered-out incompatible fns and var changes
In my initial fix for
https://sourceware.org/bugzilla/show_bug.cgi?id=31642 in commit
https://sourceware.org/cgit/libabigail/commit/?id=567d4edb880c23af9cacbe89aee7564f5c847648,
I mistakenly forgot to detect incompatible functions and variables
changes among initially filtered-out diff nodes.  A diff node can be
initially filtered out because it's deemed redundant in the initial
diff node graph walking order.  Once a diff node is later identified
as carrying an incompatible change it must not be accounted for as a
filtered diff nodes anymore.

Fixed thus.

	* include/abg-comparison.h (diff::is_categorized_as_suppressed):
	Declare ...
	* src/abg-comparison.cc (diff::is_categorized_as_suppressed):
	... new member function.
	(corpus_diff::priv::apply_filters_and_compute_diff_stats): Use the
	new diff::is_categorized_as_suppressed.  Look for incompatible fns
	and vars changes among filtered-out (and non-suppressed) diff
	nodes.
	* tests/data/test-abidiff-exit/test1-voffset-change-report0.txt:
	Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>