Bug 28505 - Fix a case of pointer to void* change filtering
Libabigail has started to categorize "void* to pointer" changes as
harmles since the patch below:

    Bug 23708 - categorize void* to pointer change as harmless

    Changing a void* pointer into another pointer of the same size is a
    change that is harmless in terms of data layout.

    This commit thus categorizes such a change as harmless.

I realized that since libabigail started to support "incompatible
changes" categorization, detecting pointer to void* changes and
classifying them as harmless started to misbehave.  Note that this is
a remnant of an ancient problem that was reported at
https://sourceware.org/bugzilla/show_bug.cgi?id=28505.

The issue with classifying pointer to void* change as harmless can be
summarized in the test below:

	$ diff -u PR28505-test-v0.c PR28505-test-v1.c
	--- PR28505-test-v0.c	2025-03-31 11:59:10.367026287 +0200
	+++ PR28505-test-v1.c	2025-03-31 11:59:17.111050102 +0200
	@@ -1,5 +1,5 @@
	-char*
	-chop(char*)
	+void*
	+chop(void*)
	 {
	   return 0;
	 }
	$

	$ abidiff PR28505-test-v0.o PR28505-test-v1.o
	Functions changes summary: 0 Removed, 1 Changed, 0 Added function
	Variables changes summary: 0 Removed, 0 Changed, 0 Added variable

	1 function with incompatible sub-type changes:
$

This is because when a pointer-to-void* change is detected, the
underlying diff node where the distinct leaf-type-to-void change
occurred usually has a NON_COMPATIBLE_DISTINCT_CHANGE_CATEGORY or a
NON_COMPATIBLE_NAME_CHANGE_CATEGORY.  When that category gets propagated to
the pointer-to-void* diff node, it gets transformed into a
VOID_PTR_TO_PTR_CHANGE_CATEGORY.

The problem is that transforming the
NON_COMPATIBLE_DISTINCT_CHANGE_CATEGORY or
NON_COMPATIBLE_NAME_CHANGE_CATEGORY into a
VOID_PTR_TO_PTR_CHANGE_CATEGORY only happens in the category of the
pointer diff node, not on its /local/ category, as it should.  Yet
that local category is used to determine the non-compatible changes on
function return types, for instance.  Note that local changes or
pointed-to-types are considered local to pointers (and reference)
types.

This patch fixes the issue by clearing the
NON_COMPATIBLE_DISTINCT_CHANGE_CATEGORY or
NON_COMPATIBLE_NAME_CHANGE_CATEGORY category from local categories of
pointer and reference diff nodes when propagating
NON_COMPATIBLE_DISTINCT_CHANGE_CATEGORY or
NON_COMPATIBLE_NAME_CHANGE_CATEGORY from their underlying diff nodes.

	* include/abg-comparison.h (print_category): Declare ...
	* src/abg-comparison.cc (print_category): ... new debugging
	function.
	(category_propagation_visitor::visit_end): Clear
	NON_COMPATIBLE_DISTINCT_CHANGE_CATEGORY
	NON_COMPATIBLE_NAME_CHANGE_CATEGORY from the local category of
	pointer/reference diff nodes when propagating those categorizing
	bits from the underlying diff nodes of pointer/reference diff
	nodes.
	* tests/data/test-abidiff-exit/PR28505-test-report.txt: New test
	reference output file.
	* tests/data/test-abidiff-exit/PR28505-test-report-2.txt:
	Likewise.
	* tests/data/test-abidiff-exit/PR28505-test-v{0,1}.{c,o}: Source
	code and binary test input files.
	* tests/data/Makefile.am: Add the new test material above to
	source distribution.
	* tests/test-abidiff-exit.cc (in_out_specs): Add the new input
	tests to this test harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>