comparison: Emit local diff node category in tree dump
The diff tree debug dump, emitted by the --dump-diff-tree option to
abidiff, does contain the category of the diff node that has been
propagated from its children nodes.  However, it does not contain the
local category of the diff node itself.

This patch adds that local category to the debug dump.

	* src/abg-comparison.cc (diff_node_printer::visit): Emit a dump of
	the local category of the current diff node.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>