tools-utils: Fix XZ decompression after all compression data is read
In xz_decompressor_type::underflow, when all compressed data is read,
the de-compressor should cycle over the already read compressed data
to decompressed it in chunks.  The current implementation gets that
wrong and quits right after all the compressed data is read and on
decompression round is completed.

This patch fixes that by making xz_decompressor_type::underflow cycle
over the already compressed data to decompress it in chunks.

	* src/abg-tools-utils.cc (xz_decompressor_type::underflow): Fix
	comment.  Decouple reading the compressed input and decompressing
	it.  Once all the compressed data is read, cycle over it to
	decompress it in chunk.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>