abg-comparison: Do not abort when a static data member is deleted
The global variable representing a static data member can appear
several times in a binary as there can be one per translation unit.
In that case, the symbols are put in a COMDAT section and the linker
merge them into once to represent its unique definition.

If that global variable is removed from the new version of the binary
being compared, it can (wrongly) appear to the diff engine as if the
global variable is being removed twice.

This patch handles that case by avoiding being overly eager in the
sanity assert for global variables detection.

	* src/abg-comparison.cc
	(corpus_diff::priv::ensure_lookup_tables_populated): Do not choke
	if a static data member is removed.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>