Bug 33264 - Add --ignore-soname to abipkgdiff
Consider two package sets (S1 and S2) where S1 contains a shared
library libfoo.so.1.0.0 which SONAME is libfoo.so.1 and S2 contains a
shared library libfoo.so.2.0.0 which SONAME is libfoo.so.2.

S1 also contains a symbolic link libfoo.so that points to
libfoo.so.1.0.0 and S2 contains a symbolic link libfoo.so that points
to libfoo.so.2.0.0.

In the current mode of operations, abipkgdiff resolves the libfoo.so
symbolic link in S1 to libfoo.so.1.0.0, resolves the libfoo.so
symbolic link in S2 to libfoo.so.2.0.0.

As libfoo.so.1.0.0 is not present in S2, abipkgdiff concludes that
libfoo.so.1.0.0 has been removed from the S2.  Similarly, it concludes
that libfoo.so.2.0.0 has been added to S2.

It can be useful to compare the ABI of the libfoo.so shared library
from S2 against the libfoo.so from S1, regardless of their SONAME
difference.

This is what a --ignore-soname would do.  And this is what this patch
implements.

With that option, the patch first forces abipkgdiff to *NOT* resolve
the symbolic link libfoo.so in both S1 and S2.  S1 and S2 are thus
considered as if they both have a regular file libfoo.so.  As they
both have a shared library libfoo.so with the same name, abipkgdiff
considers comparing their ABI.  Libabigail is thus configured to avoid
considering SONAMEs during the ABI comparison of the two libfoo.so.

	* doc/manuals/abipkgdiff.rst: Document the new option
	--ignore-soname.
	* tools/abipkgdiff.cc (options::options): Initialize ...
	(options::ignore_soname): ... the new data member.
	(parse_command_line): Parse the new --ignore-soname options ...
	(display_usage): ... and document it.
	(set_diff_context_from_opts): Set ctxt->show_soname_change
	accordingly.
	(package::convert_path_to_relative): Do not resolve symbolic links
	to their target.
	(compare_task::maybe_emit_pretty_error_message_to_output): Add
	path of the binary for which no ABI change was detected.
	(create_maps_of_package_set_content): If the user asked to ignore
	SONAMEs, then take the same path as if the current shared library
	has no SONAME.  However, no not resolve symlink paths to their
	target in that mode.
	* tests/data/test-diff-pkg/PR33264/simdjson-3.12.3-2.el10.x86_64--simdjson-3.13.0-1.el10.x86_64-report-{1,2}.txt:
	Add new reference test output.
	* tests/data/test-diff-pkg/PR33264/simdjson-3.12.3/simdjson-3.12.3-2.el10.x86_64.rpm:
	New binary test input.
	* tests/data/test-diff-pkg/PR33264/simdjson-3.12.3/simdjson-debuginfo-3.12.3-2.el10.x86_64.rpm:
	Likewise.
	* tests/data/test-diff-pkg/PR33264/simdjson-3.12.3/simdjson-devel-3.12.3-2.el10.x86_64.rpm:
	Likewise.
	* tests/data/test-diff-pkg/PR33264/simdjson-3.13.0/simdjson-3.13.0-1.el10.x86_64.rpm:
	Likewise.
	* tests/data/test-diff-pkg/PR33264/simdjson-3.13.0/simdjson-debuginfo-3.13.0-1.el10.x86_64.rpm:
	Likewise.
	* tests/data/test-diff-pkg/PR33264/simdjson-3.13.0/simdjson-devel-3.13.0-1.el10.x86_64.rpm:
	Likewise.
	* tests/data/test-diff-pkg/test-dbus-glib-0.80-3.fc12.x86_64-report-0.txt:
	Adjust
	* tests/data/test-fedabipkgdiff/vte291-0.39.1-1.fc22.x86_64--vte291-0.39.90-1.fc22.x86_64-report-0.txt:
	Adjust.
	* tests/data/Makefile.am: Add the new test material to source
	distribution.
	* tests/test-diff-pkg.cc (in_out_specs): Add the new tests inputs
	to this test harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>