dwarf-reader: Avoid adding declarations-only global vars to IR
When the DIE for a global variable doesn't have ABI, then the DWARF
reader should not create an IR for it.

That is, if the DIE for a global variable neither has a memory
location nor an ELF symbol do not add it to the IR or else, that IR
node might be the cause of spurious change reports down the road.

	* src/abg-dwarf-reader.cc (build_ir_node_from_die): When looking
	at the DIE for a variable, create an IR for it if it has an ELF
	symbol.
	* tests/data/test-annotate/libtest23.so.abi: Adjust.
	* tests/data/test-annotate/libtest24-drop-fns-2.so.abi: Likewise.
	* tests/data/test-annotate/libtest24-drop-fns.so.abi: Likewise.
	* tests/data/test-read-dwarf/libtest23.so.abi: Likewise.
	* tests/data/test-read-dwarf/libtest24-drop-fns-2.so.abi:
	Likewise.
	* tests/data/test-read-dwarf/libtest24-drop-fns.so.abi: Likewise.
	* tests/data/test-read-dwarf/test10-pr18818-gcc.so.abi: Likewise.
	* tests/data/test-read-dwarf/test11-pr18828.so.abi: Likewise.
	* tests/data/test-read-dwarf/test16-pr18904.so.abi: Likewise.
	* tests/data/test-read-dwarf/test22-pr19097-libstdc++.so.6.0.17.so.abi:
	Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>