symtab-reader: Speedup lookup_undefined_{function,variable}_symbol
Profiling showed that
symtab::lookup_undefined_{function,variable}_symbol is a hot spot in
the DWARF reader when looking at big binaries.

This function introduces a side hash map to speedup the lookup of
undefined function & variable.

On a given machine of mine, a parallel "make check" goes from 3m30s to
2m11s thanks to this change.

	* src/abg-symtab-reader.h (symtab::undefined_{fn,var}_symbols_):
	Add new data members.
	* src/abg-symtab-reader.cc (symtab::load_): Populate the newly
	introduced symtab::undefined_{fn,var}_symbols_ with undefined
	function and variable symbols.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>